began adding projectiles

+++ source/camera.c
@@ -1,4 +1,5 @@
 #include <gba.h>
+#include "gameobjects.h"
 #include "util.h"
 
 s16 camera_x = 0;
@@ -6,6 +7,6 @@ s16 camera_y = 0;
 
 void UpdateCameraOffset(GameObject* gameObject)
 {
-    camera_x = gameObject->x - (s16)(DRAW_WIDTH / 2);
-    camera_y = gameObject->y - (s16)(DRAW_HEIGHT / 2);
+    camera_x = gameObject->x - (DRAW_WIDTH / 2);
+    camera_y = gameObject->y - (DRAW_HEIGHT / 2);
 }
